Fixed issue related to Int16.Parse exceptions when Gallery ModuleId > 32767 in extremely large sites. Fix will be included in 04.03.03 RC2 submitted to release tracker.

+++ Gallery_4x/AlbumEdit.ascx.vb
@@ -443,7 +443,7 @@ Namespace DotNetNuke.Modules.Gallery
                     If sOwnerID = String.Empty Then
                         OwnerID = mGalleryConfig.OwnerID
                     Else
-                        OwnerID = Int16.Parse(sOwnerID)
+                        OwnerID = Integer.Parse(sOwnerID)
                     End If
                 Else
                     OwnerID = mUserID
